Move publish_display_data into IPython.display API
